Refactored utils.py into a subpackage
