+++ mlky/utils/__init__.py
@@ -1,0 +1,4 @@
+from .formatters  import *
+from .funcs       import *
+from .io          import *
+from .prettyslice import PrettySlice as pslice
