+++ setup.cfg
@@ -27,4 +27,4 @@ dev =
 
 [options.entry_points]
 console_scripts =
-  mlky = mlky:cli
+  mlky = mlky:CLI.group
